Remove warnings

+++ LIBRARY/src/bid64_to_uint32.c
@@ -108,7 +108,7 @@ bid64_to_uint32_rnint (BID_UINT64 x
     if (C1 >= bid_nr_digits[x_nr_bits - 1].threshold_lo)
       q++;
   }
-  exp = x_exp - 398;	// unbiased exponent
+  exp = (int) (x_exp - 398);	// unbiased exponent
 
   if ((q + exp) > 10) {	// x >= 10^10 ~= 2^33.2... (cannot fit in 32 bits)
     // set invalid flag
@@ -247,14 +247,14 @@ bid64_to_uint32_rnint (BID_UINT64 x
       Cstar--;	// Cstar is now even
     }	// else MP in [ODD, EVEN]
       }
-      res = Cstar;	// the result is positive
+      res = (unsigned int) Cstar;	// the result is positive
     } else if (exp == 0) {
       // 1 <= q <= 10
       // res = +C (exact)
-      res = C1;	// the result is positive
+      res = (unsigned int) C1;	// the result is positive
     } else {	// if (exp > 0) => 1 <= exp <= 9, 1 <= q < 9, 2 <= q + exp <= 10
       // res = +C * 10^exp (exact)
-      res = C1 * bid_ten2k64[exp];	// the result is positive
+      res = (unsigned int) (C1 * bid_ten2k64[exp]);	// the result is positive
     }
   }
   BID_RETURN (res);
@@ -339,7 +339,7 @@ bid64_to_uint32_xrnint (BID_UINT64 x
     if (C1 >= bid_nr_digits[x_nr_bits - 1].threshold_lo)
       q++;
   }
-  exp = x_exp - 398;	// unbiased exponent
+  exp = (int) (x_exp - 398);	// unbiased exponent
 
   if ((q + exp) > 10) {	// x >= 10^10 ~= 2^33.2... (cannot fit in 32 bits)
     // set invalid flag
@@ -518,14 +518,14 @@ bid64_to_uint32_xrnint (BID_UINT64 x
       Cstar--;	// Cstar is now even
     }	// else MP in [ODD, EVEN]
       }
-      res = Cstar;	// the result is positive
+      res = (unsigned int) Cstar;	// the result is positive
     } else if (exp == 0) {
       // 1 <= q <= 10
       // res = +C (exact)
-      res = C1;	// the result is positive
+      res = (unsigned int) C1;	// the result is positive
     } else {	// if (exp > 0) => 1 <= exp <= 9, 1 <= q < 9, 2 <= q + exp <= 10
       // res = +C * 10^exp (exact)
-      res = C1 * bid_ten2k64[exp];	// the result is positive
+      res = (unsigned int) (C1 * bid_ten2k64[exp]);	// the result is positive
     }
   }
   BID_RETURN (res);
@@ -616,7 +616,7 @@ bid64_to_uint32_floor (BID_UINT64 x
     if (C1 >= bid_nr_digits[x_nr_bits - 1].threshold_lo)
       q++;
   }
-  exp = x_exp - 398;	// unbiased exponent
+  exp = (int) (x_exp - 398);	// unbiased exponent
 
   if ((q + exp) > 10) {	// x >= 10^10 ~= 2^33.2... (cannot fit in 32 bits)
     // set invalid flag
@@ -698,14 +698,14 @@ bid64_to_uint32_floor (BID_UINT64 x
       shift = bid_shiftright128[ind - 1];	// 0 <= shift <= 39
       Cstar = Cstar >> shift;
 
-      res = Cstar;	// the result is positive
+      res = (unsigned int) Cstar;	// the result is positive
     } else if (exp == 0) {
       // 1 <= q <= 10
       // res = +C (exact)
-      res = C1;	// the result is positive
+      res = (unsigned int) C1;	// the result is positive
     } else {	// if (exp > 0) => 1 <= exp <= 9, 1 <= q < 9, 2 <= q + exp <= 10
       // res = +C * 10^exp (exact)
-      res = C1 * bid_ten2k64[exp];	// the result is positive
+      res = (unsigned int) (C1 * bid_ten2k64[exp]);	// the result is positive
     }
   }
   BID_RETURN (res);
@@ -797,7 +797,7 @@ bid64_to_uint32_xfloor (BID_UINT64 x
     if (C1 >= bid_nr_digits[x_nr_bits - 1].threshold_lo)
       q++;
   }
-  exp = x_exp - 398;	// unbiased exponent
+  exp = (int) (x_exp - 398);	// unbiased exponent
 
   if ((q + exp) > 10) {	// x >= 10^10 ~= 2^33.2... (cannot fit in 32 bits)
     // set invalid flag
@@ -903,14 +903,14 @@ bid64_to_uint32_xfloor (BID_UINT64 x
     }	// else the result is exact
       }
 
-      res = Cstar;	// the result is positive
+      res = (unsigned int) Cstar;	// the result is positive
     } else if (exp == 0) {
       // 1 <= q <= 10
       // res = +C (exact)
-      res = C1;	// the result is positive
+      res = (unsigned int) C1;	// the result is positive
     } else {	// if (exp > 0) => 1 <= exp <= 9, 1 <= q < 9, 2 <= q + exp <= 10
       // res = +C * 10^exp (exact)
-      res = C1 * bid_ten2k64[exp];	// the result is positive
+      res = (unsigned int) (C1 * bid_ten2k64[exp]);	// the result is positive
     }
   }
   BID_RETURN (res);
@@ -995,7 +995,7 @@ bid64_to_uint32_ceil (BID_UINT64 x
     if (C1 >= bid_nr_digits[x_nr_bits - 1].threshold_lo)
       q++;
   }
-  exp = x_exp - 398;	// unbiased exponent
+  exp = (int) (x_exp - 398);	// unbiased exponent
 
   if ((q + exp) > 10) {	// x >= 10^10 ~= 2^33.2... (cannot fit in 32 bits)
     // set invalid flag
@@ -1116,14 +1116,14 @@ bid64_to_uint32_ceil (BID_UINT64 x
     }	// else the result is exact
       }
 
-      res = Cstar;	// the result is positive
+      res = (unsigned int) Cstar;	// the result is positive
     } else if (exp == 0) {
       // 1 <= q <= 10
       // res = +C (exact)
-      res = C1;	// the result is positive
+      res = (unsigned int) C1;	// the result is positive
     } else {	// if (exp > 0) => 1 <= exp <= 9, 1 <= q < 9, 2 <= q + exp <= 10
       // res = +C * 10^exp (exact)
-      res = C1 * bid_ten2k64[exp];	// the result is positive
+      res = (unsigned int) (C1 * bid_ten2k64[exp]);	// the result is positive
     }
   }
   BID_RETURN (res);
@@ -1208,7 +1208,7 @@ bid64_to_uint32_xceil (BID_UINT64 x
     if (C1 >= bid_nr_digits[x_nr_bits - 1].threshold_lo)
       q++;
   }
-  exp = x_exp - 398;	// unbiased exponent
+  exp = (int) (x_exp - 398);	// unbiased exponent
 
   if ((q + exp) > 10) {	// x >= 10^10 ~= 2^33.2... (cannot fit in 32 bits)
     // set invalid flag
@@ -1335,14 +1335,14 @@ bid64_to_uint32_xceil (BID_UINT64 x
     }	// else the result is exact
       }
 
-      res = Cstar;	// the result is positive
+      res = (unsigned int) Cstar;	// the result is positive
     } else if (exp == 0) {
       // 1 <= q <= 10
       // res = +C (exact)
-      res = C1;	// the result is positive
+      res = (unsigned int) C1;	// the result is positive
     } else {	// if (exp > 0) => 1 <= exp <= 9, 1 <= q < 9, 2 <= q + exp <= 10
       // res = +C * 10^exp (exact)
-      res = C1 * bid_ten2k64[exp];	// the result is positive
+      res = (unsigned int) (C1 * bid_ten2k64[exp]);	// the result is positive
     }
   }
   BID_RETURN (res);
@@ -1426,7 +1426,7 @@ bid64_to_uint32_int (BID_UINT64 x
     if (C1 >= bid_nr_digits[x_nr_bits - 1].threshold_lo)
       q++;
   }
-  exp = x_exp - 398;	// unbiased exponent
+  exp = (int) (x_exp - 398);	// unbiased exponent
 
   if ((q + exp) > 10) {	// x >= 10^10 ~= 2^33.2... (cannot fit in 32 bits)
     // set invalid flag
@@ -1524,14 +1524,14 @@ bid64_to_uint32_int (BID_UINT64 x
       shift = bid_shiftright128[ind - 1];	// 0 <= shift <= 39
       Cstar = Cstar >> shift;
 
-      res = Cstar;	// the result is positive
+      res = (unsigned int) Cstar;	// the result is positive
     } else if (exp == 0) {
       // 1 <= q <= 10
       // res = +C (exact)
-      res = C1;	// the result is positive
+      res = (unsigned int) C1;	// the result is positive
     } else {	// if (exp > 0) => 1 <= exp <= 9, 1 <= q < 9, 2 <= q + exp <= 10
       // res = +C * 10^exp (exact)
-      res = C1 * bid_ten2k64[exp];	// the result is positive
+      res = (unsigned int) (C1 * bid_ten2k64[exp]);	// the result is positive
     }
   }
   BID_RETURN (res);
@@ -1616,7 +1616,7 @@ bid64_to_uint32_xint (BID_UINT64 x
     if (C1 >= bid_nr_digits[x_nr_bits - 1].threshold_lo)
       q++;
   }
-  exp = x_exp - 398;	// unbiased exponent
+  exp = (int) (x_exp - 398);	// unbiased exponent
 
   if ((q + exp) > 10) {	// x >= 10^10 ~= 2^33.2... (cannot fit in 32 bits)
     // set invalid flag
@@ -1738,14 +1738,14 @@ bid64_to_uint32_xint (BID_UINT64 x
     }	// else the result is exact
       }
 
-      res = Cstar;	// the result is positive
+      res = (unsigned int) Cstar;	// the result is positive
     } else if (exp == 0) {
       // 1 <= q <= 10
       // res = +C (exact)
-      res = C1;	// the result is positive
+      res = (unsigned int) C1;	// the result is positive
     } else {	// if (exp > 0) => 1 <= exp <= 9, 1 <= q < 9, 2 <= q + exp <= 10
       // res = +C * 10^exp (exact)
-      res = C1 * bid_ten2k64[exp];	// the result is positive
+      res = (unsigned int) (C1 * bid_ten2k64[exp]);	// the result is positive
     }
   }
   BID_RETURN (res);
@@ -1829,7 +1829,7 @@ bid64_to_uint32_rninta (BID_UINT64 x
     if (C1 >= bid_nr_digits[x_nr_bits - 1].threshold_lo)
       q++;
   }
-  exp = x_exp - 398;	// unbiased exponent
+  exp = (int) (x_exp - 398);	// unbiased exponent
 
   if ((q + exp) > 10) {	// x >= 10^10 ~= 2^33.2... (cannot fit in 32 bits)
     // set invalid flag
@@ -1948,14 +1948,14 @@ bid64_to_uint32_rninta (BID_UINT64 x
       Cstar = Cstar >> shift;
 
       // if the result was a midpoint it was rounded away from zero
-      res = Cstar;	// the result is positive
+      res = (unsigned int) Cstar;	// the result is positive
     } else if (exp == 0) {
       // 1 <= q <= 10
       // res = +C (exact)
-      res = C1;	// the result is positive
+      res = (unsigned int) C1;	// the result is positive
     } else {	// if (exp > 0) => 1 <= exp <= 9, 1 <= q < 9, 2 <= q + exp <= 10
       // res = +C * 10^exp (exact)
-      res = C1 * bid_ten2k64[exp];	// the result is positive
+      res = (unsigned int) (C1 * bid_ten2k64[exp]);	// the result is positive
     }
   }
   BID_RETURN (res);
@@ -2040,7 +2040,7 @@ bid64_to_uint32_xrninta (BID_UINT64 x
     if (C1 >= bid_nr_digits[x_nr_bits - 1].threshold_lo)
       q++;
   }
-  exp = x_exp - 398;	// unbiased exponent
+  exp = (int) (x_exp - 398);	// unbiased exponent
 
   if ((q + exp) > 10) {	// x >= 10^10 ~= 2^33.2... (cannot fit in 32 bits)
     // set invalid flag
@@ -2202,14 +2202,14 @@ bid64_to_uint32_xrninta (BID_UINT64 x
       }
 
       // if the result was a midpoint it was rounded away from zero
-      res = Cstar;	// the result is positive
+      res = (unsigned int) Cstar;	// the result is positive
     } else if (exp == 0) {
       // 1 <= q <= 10
       // res = +C (exact)
-      res = C1;	// the result is positive
+      res = (unsigned int) C1;	// the result is positive
     } else {	// if (exp > 0) => 1 <= exp <= 9, 1 <= q < 9, 2 <= q + exp <= 10
       // res = +C * 10^exp (exact)
-      res = C1 * bid_ten2k64[exp];	// the result is positive
+      res = (unsigned int) (C1 * bid_ten2k64[exp]);	// the result is positive
     }
   }
   BID_RETURN (res);
